requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0 